switched from sveltekit to svelte; using newer version of threlte; nothing broke :)
