+++ frontend/svelte.config.js
@@ -1,20 +1,7 @@
-import adapter from '@sveltejs/adapter-auto';
-import preprocess from 'svelte-preprocess';
+import sveltePreprocess from 'svelte-preprocess'
 
-/** @type {import('@sveltejs/kit').Config} */
-const config = {
-	// Consult https://github.com/sveltejs/svelte-preprocess
-	// for more information about preprocessors
-	preprocess: preprocess(),
-
-	kit: {
-		adapter: adapter(),
-		vite: {
-			ssr: {
-				noExternal: ['three']
-			}
-		}
-	}
-};
-
-export default config;
+export default {
+  // Consult https://github.com/sveltejs/svelte-preprocess
+  // for more information about preprocessors
+  preprocess: sveltePreprocess()
+}
